fixed windows build breaks

+++ inc/rapid-vulkan/rapid-vulkan.h
@@ -1462,19 +1462,19 @@ public:
 
         template<typename T, size_t C>
         ConstructParameters & setSpirv(const std::array<T, C> & data) {
-            spirv = vk::ArrayProxy<const uint32_t>(C * sizeof(T) / sizeof(uint32_t), (const uint32_t *) data.data());
+            spirv = vk::ArrayProxy<const uint32_t>((uint32_t) (C * sizeof(T) / sizeof(uint32_t)), (const uint32_t *) data.data());
             return *this;
         }
 
         template<typename T, size_t C>
         ConstructParameters & setSpirv(const T (&data)[C]) {
-            spirv = vk::ArrayProxy<const uint32_t>(C * sizeof(T) / sizeof(uint32_t), (const uint32_t *) data);
+            spirv = vk::ArrayProxy<const uint32_t>((uint32_t) (C * sizeof(T) / sizeof(uint32_t)), (const uint32_t *) data);
             return *this;
         }
 
         template<typename T>
         ConstructParameters & setSpirv(size_t countInUnitOfT, const T * data) {
-            spirv = vk::ArrayProxy<const uint32_t>(countInUnitOfT * sizeof(T) / sizeof(uint32_t), (const uint32_t *) data);
+            spirv = vk::ArrayProxy<const uint32_t>((uint32_t) (countInUnitOfT * sizeof(T) / sizeof(uint32_t)), (const uint32_t *) data);
             return *this;
         }
